Converted Entity to a class w/ convenient methods
Added a variety of methods to make referencing the World obsolete.
So far, this has proved to be a much simplier way for the user to
create, change, and remove Components from an Entity.

This way of creating, changing, and removing Components from the
Entity itself just seems more natural to program in.