Added versioning. Welcome to Divvy v0.5!
I'm proud to say that Divvy has reached enough stability to version and
document. For the bulk of v0.5, I've focused on documentation and
releasing it for other eyes. I have improved the way methods are
commented, adding more formality and ease. To make it official, I have
even implemented the move constructor and operator for Entity.

I know I got the bulk of the Component Entity done, but there is still
much to do. That is why I started the versioning at v0.5 instead of
starting at v1.0.

Some future additions that I look forward to adding:
    - Messaging System
    - Benchmarking

Again, welcome to Divvy v0.5!